backend/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy
pandas
python-multipart
websockets
sqlalchemy
librosa
soundfile
groq
deep-sort-realtime
python-dotenv
